+++ pkg/webui/console/store/actions/applications.js
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { APPLICATION } from '@console/constants/entities'
+
 import createRequestActions from '@ttn-lw/lib/store/actions/create-request-actions'
 import {
   createPaginationRequestActions,
@@ -43,7 +45,7 @@ import {
   setEventsFilter,
 } from './events'
 
-export const SHARED_NAME = 'APPLICATION'
+export const SHARED_NAME = APPLICATION
 
 export const GET_APP_BASE = 'GET_APPLICATION'
 export const [