+++ pkg/webui/console/store/actions/devices.js
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import { END_DEVICE } from '@console/constants/entities'
+
 import createRequestActions from '@ttn-lw/lib/store/actions/create-request-actions'
 import {
   createPaginationByIdRequestActions,
@@ -37,9 +39,9 @@ import {
   setEventsFilter,
 } from './events'
 
-export const SHARED_NAME = 'DEVICE'
+export const SHARED_NAME = END_DEVICE
 
-export const GET_DEV_BASE = 'GET_DEVICE'
+export const GET_DEV_BASE = 'GET_END_DEVICE'
 export const [
   { request: GET_DEV, success: GET_DEV_SUCCESS, failure: GET_DEV_FAILURE },
   { request: getDevice, success: getDeviceSuccess, failure: getDeviceFailure },
@@ -49,7 +51,7 @@ export const [
   (appId, deviceId, selector, options) => ({ selector, options }),
 )
 
-export const UPDATE_DEV_BASE = 'UPDATE_DEVICE'
+export const UPDATE_DEV_BASE = 'UPDATE_END_DEVICE'
 export const [
   { request: UPDATE_DEV, success: UPDATE_DEV_SUCCESS, failure: UPDATE_DEV_FAILURE },
   { request: updateDevice, success: updateDeviceSuccess, failure: updateDeviceFailure },